+++ swing/framework-model/src/main/java/is/codion/swing/framework/model/SwingEntityTableModel.java
@@ -1087,9 +1087,13 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
       return attributes.get();
     }
 
-    return attributes.get().stream()
-            .filter(attribute -> columnModel.visible(attribute).get())
+    return entityDefinition().attributes().selected().stream()
+            .filter(this::columnNotHidden)
             .collect(toList());
+  }
+
+  private boolean columnNotHidden(Attribute<?> attribute) {
+    return !columnModel().containsColumn(attribute) || columnModel().visible(attribute).get();
   }
 
   private class AttributeValidator implements Value.Validator<Set<Attribute<?>>> {
@@ -1204,18 +1208,14 @@ public class SwingEntityTableModel implements EntityTableModel<SwingEntityEditMo
 
     @Override
     public Optional<ColumnConditionModel<? extends Attribute<?>, ?>> createConditionModel(Attribute<?> attribute) {
-      if (requireNonNull(attribute).type().isEntity()) {
+      AttributeDefinition<?> attributeDefinition = entityDefinition.attributes().definition(attribute);
+      if (attributeDefinition.hidden() || !Comparable.class.isAssignableFrom(attribute.type().valueClass())) {
+        return Optional.empty();
+      }
+      if (requireNonNull(attribute).type().isEntity() || !attributeDefinition.items().isEmpty()) {
         return Optional.of(ColumnConditionModel.builder(attribute, String.class)
                 .operators(operators(String.class))
                 .build());
-      }
-      if (!Comparable.class.isAssignableFrom(attribute.type().valueClass())) {
-        return Optional.empty();
-      }
-
-      AttributeDefinition<?> attributeDefinition = entityDefinition.attributes().definition(attribute);
-      if (attributeDefinition.hidden()) {
-        return Optional.empty();
       }
 
       return Optional.of(ColumnConditionModel.builder(attribute, attribute.type().valueClass())