+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/EntityComboBoxModelTest.java
@@ -95,6 +95,8 @@ public final class EntityComboBoxModelTest {
 
     EntityEditEvents.notifyInserted(singletonList(temp));
     assertFalse(comboBoxModel.visible(temp));
+
+    comboBoxModel.refresh();
   }
 
   @Test